Many fields moved from Main to MainLoader
The rest of the application should not know anything about Main, so Main should not have any public fields. They have been moved to the class where they are written: MainLoader.

Getter methods have been introduced in MainLoader, and the (moved) fields have been made private.

MainLoader is not likely to be the final destination of all these fields.